% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin_key.R
\name{begin_key}
\alias{begin_key}
\title{Generate the Skeleton of a Column Key}
\usage{
begin_key(
  raw_folder = NULL,
  data_format = c("csv", "txt", "xls", "xlsx"),
  guess_tidy = FALSE
)
}
\arguments{
\item{raw_folder}{(character) folder / folder path containing data files to include in key}

\item{data_format}{(character) file extensions to identify within the \code{raw_folder}. Default behavior is to search for all supported file types.}

\item{guess_tidy}{(logical) whether to attempt to "guess" what the tidy name equivalent should be for each raw column name. This is accomplished via coercion to lowercase and removal of special character/repeated characters. If \code{FALSE} (the default) the "tidy_name" column is returned empty}
}
\value{
(dataframe) skeleton of column key
}
\description{
Creates the start of a 'column key' for harmonizing data. A column key includes a column for the file names to be harmonized into a single data object as well as a column for the column names in those files. Finally, it includes a column indicating the tidied name that corresponds with each raw column name. Harmonization can accept this key object and use it to rename all raw column names--in a reproducible way--to standardize across datasets. Currently supports raw files of the following formats: CSV, TXT, XLS, and XLSX
}
\examples{
\dontrun{
# Generate two simple tables
## Dataframe 1
df1 <- data.frame("xx" = c(1:3),
                  "unwanted" = c("not", "needed", "column"),
                  "yy" = letters[1:3])
## Dataframe 2
df2 <- data.frame("LETTERS" = letters[4:7],
                  "NUMBERS" = c(4:7),
                  "BONUS" = c("plantae", "animalia", "fungi", "protista"))

# Generate a local folder for exporting
dir.create(path = "ltertools_test", showWarnings = FALSE)

# Export both files to that folder
utils::write.csv(x = df1, file = file.path("ltertools_test", "df1.csv"), row.names = FALSE)
utils::write.csv(x = df2, file = file.path("ltertools_test", "df2.csv"), row.names = FALSE)

# Generate a column key with "guesses" at tidy column names
ltertools::begin_key(raw_folder = "ltertools_test", data_format = "csv", guess_tidy = TRUE)
}

}
