% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_temp}
\alias{convert_temp}
\title{Convert Temperature Values}
\usage{
convert_temp(value = NULL, from = NULL, to = NULL)
}
\arguments{
\item{value}{(numeric) temperature values to convert}

\item{from}{(character) starting units of the value, not case sensitive.}

\item{to}{(character) units to which to convert, not case sensitive.}
}
\value{
(numeric) converted temperature values
}
\description{
Converts a given set of temperature values from one unit to another. This function was built by the following authors: Nicholas Lyon
}
\examples{
# Convert from Fahrenheit to Celsius
convert_temp(value = 32, from = "Fahrenheit", to = "c")

}
