% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltertools-package.R
\docType{package}
\name{ltertools-package}
\alias{ltertools}
\alias{ltertools-package}
\title{ltertools: Tools Developed by the Long Term Ecological Research Community}
\description{
Set of the data science tools created by various members of the Long Term Ecological Research (LTER) community. These functions were initially written largely as standalone operations and have later been aggregated into this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lter.github.io/ltertools/}
  \item Report bugs at \url{https://github.com/lter/ltertools/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Lyon \email{lyon@nceas.ucsb.edu} (https://njlyon0.github.io/)

Authors:
\itemize{
  \item Nicholas Lyon \email{lyon@nceas.ucsb.edu} (https://njlyon0.github.io/)
  \item Angel Chen \email{anchen@nceas.ucsb.edu} (https://angelchen7.github.io)
}

Other contributors:
\itemize{
  \item Miguel C. Leon (https://luquillo.lter.network/) [contributor]
  \item  National Science Foundation (NSF 1929393, 09/01/2019 - 08/31/2024) [funder]
  \item  University of California, Santa Barbara [copyright holder]
}

}
\keyword{internal}
