% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read Data from Folder}
\usage{
read(raw_folder = NULL, data_format = c("csv", "txt", "xls", "xlsx"))
}
\arguments{
\item{raw_folder}{(character) folder / folder path containing data files to read}

\item{data_format}{(character) file extensions to identify within the \code{raw_folder}. Default behavior is to search for all supported file types.}
}
\value{
(list) data found in specified folder of specified file format(s)
}
\description{
Reads in all data files of specified types found in the designated folder. Returns a list with one element for each data file. Currently supports CSV, TXT, XLS, and XLSX. This function was built by the following authors: Nicholas Lyon
}
\examples{
\dontrun{
# Generate two simple tables
## Dataframe 1
df1 <- data.frame("xx" = c(1:3),
                  "unwanted" = c("not", "needed", "column"),
                  "yy" = letters[1:3])
## Dataframe 2
df2 <- data.frame("LETTERS" = letters[4:7],
                  "NUMBERS" = c(4:7),
                  "BONUS" = c("plantae", "animalia", "fungi", "protista"))

# Generate a local folder for exporting
dir.create(path = "ltertools_test", showWarnings = FALSE)

# Export both files to that folder
utils::write.csv(x = df1, file = file.path("ltertools_test", "df1.csv"), row.names = FALSE)
utils::write.csv(x = df2, file = file.path("ltertools_test", "df2.csv"), row.names = FALSE)

# Read in all CSV files in that folder
read(raw_folder = "ltertools_test", data_format = "csv")
}

}
