% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_subset.R
\name{site_subset}
\alias{site_subset}
\title{Subsets the LTER Site Information Table by Site Codes and Habitats}
\usage{
site_subset(sites = NULL, habitats = NULL)
}
\arguments{
\item{sites}{(character) three letter site code(s) identifying site(s) of interest}

\item{habitats}{(character) habitat(s) of interest. See `unique(lter_sites`}
}
\value{
(dataframe) complete site information (8 columns) for all sites that meet the provided site code and/or habitat criteria
}
\description{
Subsets the information on long term ecological research (LTER) sites based on user-specified site codes (i.e., three letter abbreviations), and/or desired habitats. See [lter_sites] for the full set of site information. This function was built by the following authors: Nicholas Lyon
}
\examples{
\dontrun{
# See only marine or coastal sites
site_subset(habitats = c("marine", "coastal"))

# Return only specified sites
site_subset(sites = c("AND", "NGA", "BLE", "LNO"))
}

}
