% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltertools-package.R
\docType{package}
\name{ltertools-package}
\alias{ltertools}
\alias{ltertools-package}
\title{ltertools: Tools Developed by the Long Term Ecological Research Community}
\description{
Set of the data science tools created by various members of the Long Term Ecological Research (LTER) community. These functions were initially written largely as standalone operations and have later been aggregated into this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lter.github.io/ltertools/}
  \item Report bugs at \url{https://github.com/lter/ltertools/issues}
}

}
\keyword{internal}
